Branch: MatrixPilot_IP - Merged all trunk changes and includes new features from _NASA branch. - Biggest additions from trunk were stdint.h definitions, SILSIM and HILSIM enhancements - This branch is officially stale. All further IP research will be done on _NASA branch
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_IP@2182 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/pitchCntrl.c
@@ -26,31 +26,31 @@
 
 #define ANGLE_90DEG (RMAX/(2*57.3))
 
-#define RTLKICK ((long)(RTL_PITCH_DOWN*(RMAX/57.3)))
-#define INVNPITCH ((long)(INVERTED_NEUTRAL_PITCH*(RMAX/57.3)))
-#define HOVERPOFFSET ((long)(HOVER_PITCH_OFFSET*(RMAX/57.3)))
-#define HOVERPTOWP ((long)(HOVER_PITCH_TOWARDS_WP*(RMAX/57.3)))
+#define RTLKICK ((int32_t)(RTL_PITCH_DOWN*(RMAX/57.3)))
+#define INVNPITCH ((int32_t)(INVERTED_NEUTRAL_PITCH*(RMAX/57.3)))
+#define HOVERPOFFSET ((int32_t)(HOVER_PITCH_OFFSET*(RMAX/57.3)))
+#define HOVERPTOWP ((int32_t)(HOVER_PITCH_TOWARDS_WP*(RMAX/57.3)))
 
 #if ((SERIAL_OUTPUT_FORMAT == SERIAL_MAVLINK) || ( GAINS_VARIABLE == 1 ))
-	int pitchgain = (int)(PITCHGAIN*RMAX) ;
-	int pitchkd = (int) (PITCHKD*SCALEGYRO*RMAX) ;
-	int hoverpitchgain = (int)(HOVER_PITCHGAIN*RMAX) ;
-	int hoverpitchkd = (int) (HOVER_PITCHKD*SCALEGYRO*RMAX) ;
-	int rudderElevMixGain = (int)(RMAX*RUDDER_ELEV_MIX) ;
-	int rollElevMixGain = (int)(RMAX*ROLL_ELEV_MIX) ;
+	int16_t pitchgain = (int16_t)(PITCHGAIN*RMAX) ;
+	int16_t pitchkd = (int16_t) (PITCHKD*SCALEGYRO*RMAX) ;
+	int16_t hoverpitchgain = (int16_t)(HOVER_PITCHGAIN*RMAX) ;
+	int16_t hoverpitchkd = (int16_t) (HOVER_PITCHKD*SCALEGYRO*RMAX) ;
+	int16_t rudderElevMixGain = (int16_t)(RMAX*RUDDER_ELEV_MIX) ;
+	int16_t rollElevMixGain = (int16_t)(RMAX*ROLL_ELEV_MIX) ;
 #else
-	const int pitchgain = (int)(PITCHGAIN*RMAX) ;
-	const int pitchkd = (int) (PITCHKD*SCALEGYRO*RMAX) ;
-	const int hoverpitchgain = (int)(HOVER_PITCHGAIN*RMAX) ;
-	const int hoverpitchkd = (int) (HOVER_PITCHKD*SCALEGYRO*RMAX) ;
-	const int rudderElevMixGain = (int)(RMAX*RUDDER_ELEV_MIX) ;
-	const int rollElevMixGain = (int)(RMAX*ROLL_ELEV_MIX) ;
+	const int16_t pitchgain = (int16_t)(PITCHGAIN*RMAX) ;
+	const int16_t pitchkd = (int16_t) (PITCHKD*SCALEGYRO*RMAX) ;
+	const int16_t hoverpitchgain = (int16_t)(HOVER_PITCHGAIN*RMAX) ;
+	const int16_t hoverpitchkd = (int16_t) (HOVER_PITCHKD*SCALEGYRO*RMAX) ;
+	const int16_t rudderElevMixGain = (int16_t)(RMAX*RUDDER_ELEV_MIX) ;
+	const int16_t rollElevMixGain = (int16_t)(RMAX*ROLL_ELEV_MIX) ;
 #endif
 
 
-int pitchrate ;
-int navElevMix ;
-int elevInput ;
+int16_t pitchrate ;
+int16_t navElevMix ;
+int16_t elevInput ;
 
 void normalPitchCntrl(void) ;
 void hoverPitchCntrl(void) ;
@@ -73,8 +73,8 @@ void pitchCntrl(void)
 void normalPitchCntrl(void)
 {
 	union longww pitchAccum ;
-	int rtlkick ;
-//	int aspd_adj ;
+	int16_t rtlkick ;
+//	int16_t aspd_adj ;
 //	fractional aspd_err, aspd_diff ;
 	
 #ifdef TestGains
@@ -147,7 +147,7 @@ void normalPitchCntrl(void)
 		pitchAccum.WW = 0 ;
 	}
 	
-	pitch_control = (long)pitchAccum._.W1 + navElevMix ;
+	pitch_control = (int32_t)pitchAccum._.W1 + navElevMix ;
 
 	return ;
 }
@@ -164,10 +164,10 @@ void hoverPitchCntrl(void)
 						- __builtin_mulss( rmat[6] , omegagyro[1] )) << 1 ;
 		pitchrate = pitchAccum._.W1 ;
 		
-		int elevInput = ( udb_flags._.radio_on == 1 ) ? REVERSE_IF_NEEDED(ELEVATOR_CHANNEL_REVERSED, udb_pwIn[ELEVATOR_INPUT_CHANNEL] - udb_pwTrim[ELEVATOR_INPUT_CHANNEL]) : 0 ;
-		int manualPitchOffset = elevInput * (int)(RMAX/600);
+		int16_t elevInput = ( udb_flags._.radio_on == 1 ) ? REVERSE_IF_NEEDED(ELEVATOR_CHANNEL_REVERSED, udb_pwIn[ELEVATOR_INPUT_CHANNEL] - udb_pwTrim[ELEVATOR_INPUT_CHANNEL]) : 0 ;
+		int16_t manualPitchOffset = elevInput * (int16_t)(RMAX/600);
 		
-		long pitchToWP ;
+		int32_t pitchToWP ;
 		
 		if ( flags._.GPS_steering )
 		{
@@ -186,7 +186,7 @@ void hoverPitchCntrl(void)
 		pitchAccum.WW = 0 ;
 	}
 	
-	pitch_control = (long)pitchAccum._.W1 ;
+	pitch_control = (int32_t)pitchAccum._.W1 ;
 	
 	return ;
 }
